pc_recommendation/requirements.txt
Flask==2.1.3
requests==2.28.1
google-cloud-dialogflow==2.15.0
python-dotenv==0.20.0
pandas==1.3.5
